web_requirements.txt
Flask==2.3.3
Pillow>=8.0.0
click>=8.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
gunicorn==21.2.0
